Add files via upload
The top panel shows the frequency of ride-hailing pickups during the week preceding 12/16/2018 at two locations with coordinates: (41.88099447, -87.63274649) and (41.89204214, -87.63186395). The top graph shows a general trend of peaks and troughs according to a day/night schedule as well as a shift in the latter location experiencing more pickups at a later time than the former location; moreover, the absolute frequency of the latter location increase as the date approaches the weekend where the former location remains more constant.
The bottom panel shows a specific day of data for location (41.88099447, -87.63274649), where the profile has been fit to multiple gaussian functions. The locations of the gaussian features suggest peaks around 1PM, 6PM, 10PM, and 1PM.